Utility Suite/requirements.txt
requests
psutil
plyer
python-dotenv
pynput
